connected clients reporting and code clean up
To find connected clients the TVHEADEND_IP was used. This does not work if using localhost as IP even though the API of tvheadend can be accessed using this IP. Using localhost as IP is beneficial because this way one can restrict the network of the "autosuspend" user. It is sufficient to detect connections just using the port number. In addition the connected client is now reported using clear names.

The code for queryTvheadend is slightly cleaned up and now also reports the raw error code of the curl query.